+++ src/main/java/roomescape/config/WebMvcConfig.java
@@ -5,6 +5,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.HandlerInterceptor;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 import java.util.List;
@@ -21,6 +22,12 @@ public class WebMvcConfig implements WebMvcConfigurer {
     ) {
         this.loginMemberArgumentResolver = loginMemberArgumentResolver;
         this.roleCheckHandlerInterceptor = roleCheckHandlerInterceptor;
+    }
+
+    @Override
+    public void addViewControllers(ViewControllerRegistry registry) {
+        registry.addViewController("/reservation")
+                .setViewName("reservation");
     }
 
     @Override
