+++ src/main/java/roomescape/reservation/controller/AdminReservationController.java
@@ -1,0 +1,46 @@
+package roomescape.reservation.controller;
+
+import jakarta.validation.Valid;
+import org.springframework.web.bind.annotation.*;
+import roomescape.reservation.dto.ReservationCreateRequest;
+import roomescape.reservation.dto.ReservationResponse;
+import roomescape.reservation.service.ReservationService;
+
+import java.time.LocalDate;
+import java.util.List;
+
+@RestController
+@RequestMapping("/admin/reservations")
+public class AdminReservationController {
+
+    private final ReservationService reservationService;
+
+    public AdminReservationController(ReservationService reservationService) {
+        this.reservationService = reservationService;
+    }
+
+    @PostMapping
+    public ReservationResponse createReservation(@Valid @RequestBody ReservationCreateRequest request) {
+        return reservationService.createReservation(request);
+    }
+
+    @GetMapping
+    public List<ReservationResponse> readReservations() {
+        return reservationService.readReservations();
+    }
+
+    @GetMapping("/search")
+    public List<ReservationResponse> readReservations(
+            @RequestParam LocalDate dateFrom,
+            @RequestParam LocalDate dateTo,
+            @RequestParam Long memberId,
+            @RequestParam Long themeId
+    ) {
+        return reservationService.searchReservations(dateFrom, dateTo, memberId, themeId);
+    }
+
+    @DeleteMapping("/{id}")
+    public void deleteReservation(@PathVariable Long id) {
+        reservationService.deleteReservation(id);
+    }
+}
