--- src/main/java/roomescape/reservation/controller/UserReservationController.java
@@ -1,15 +1,0 @@
-package roomescape.reservation.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("/reservation")
-public class UserReservationController {
-
-    @GetMapping
-    public String readUserReservation() {
-        return "reservation";
-    }
-}
